+++ lib/jazzy/config.rb
@@ -22,6 +22,7 @@ module Jazzy
     attr_accessor :root_url
     attr_accessor :version
     attr_accessor :min_acl
+    attr_accessor :skip_undocumented
 
     def initialize
       self.output = Pathname('docs')
@@ -33,6 +34,7 @@ module Jazzy
       self.docset_platform = 'jazzy'
       self.version = '1.0'
       self.min_acl = SourceDeclaration::AccessControlLevel.internal
+      self.skip_undocumented = false
     end
 
     # rubocop:disable Metrics/MethodLength
@@ -119,6 +121,13 @@ module Jazzy
           elsif acl == 'public'
             config.min_acl = SourceDeclaration::AccessControlLevel.public
           end
+        end
+
+        opt.on('--[no-]skip-undocumented',
+               "Don't document declarations that have no documentation \
+               comments.",
+               ) do |skip_undocumented|
+          config.skip_undocumented = skip_undocumented
         end
 
         opt.on('-v', '--version', 'Print version number') do