+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ##### Breaking
 
-None.
+* Don't skip declarations with no documentation comments by default.
+  Allow skipping using `--skip-undocumented`.  
+  [JP Simard](https://github.com/jpsim)
+  [#129](https://github.com/realm/jazzy/issues/129)
 
 ##### Enhancements